+++ tests/test_ellipsoids.py
@@ -1,5 +1,5 @@
 """
-This Python 3.3 code tests the ``ellipsoids`` module.
+This Python 3.11 code tests the ``ellipsoids`` module.
 Beware, these tests cover only some functions and only some scenarios.
 Keep adding tests!